requirements.txt
boto3==1.20.24
gevent==21.12.0
paramiko==2.10.1
python-daemon==2.3.0
requests==2.26.0
schedule==1.1.0
azure-identity==1.7.0
azure-mgmt-compute==21.0.0
azure-mgmt-network==19.0.0